* added combined color pipe abstraction * added further grouping support * this broke a few tests and the example but will be fixed in later versions

+++ heatmap-core/src/main/java/de/uniwue/dmir/heatmap/core/visualizer/IAlphaScheme.java
@@ -22,5 +22,5 @@ package de.uniwue.dmir.heatmap.core.visualizer;
 
 
 public interface IAlphaScheme {
-	double getColor(int value);
+	int getColor(double value);
 }
